+++ src/Commands/Self/ClearCacheCommand.php
@@ -5,6 +5,7 @@ namespace Pantheon\Terminus\Commands\Self;
 use League\Container\ContainerAwareInterface;
 use League\Container\ContainerAwareTrait;
 use Pantheon\Terminus\Commands\TerminusCommand;
+use Pantheon\Terminus\Helpers\LocalMachineHelper;
 
 /**
  * Class ClearCacheCommand
@@ -24,6 +25,14 @@ class ClearCacheCommand extends TerminusCommand implements ContainerAwareInterfa
      */
     public function clearCache()
     {
+        $local_machine = $this->getContainer()->get(LocalMachineHelper::class);
+        $fs = $local_machine->getFilesystem();
+        $finder = $local_machine->getFinder();
+
+        $finder->files()->in($this->getConfig()->get('command_cache_dir'));
+        foreach ($finder as $file) {
+            $fs->remove($file);
+        }
         $this->log()->notice('The local Terminus cache has been cleared.');
     }
 }